gateway events for audit log and auto moderation
