+++ discord-haskell.cabal
@@ -161,6 +161,7 @@ library
     , Discord.Internal.Types.RolePermissions
     , Discord.Internal.Types.ScheduledEvents
     , Discord.Internal.Types.AutoModeration
+    , Discord.Internal.Types.AuditLog
   build-depends:
   -- https://gitlab.haskell.org/ghc/ghc/-/wikis/commentary/libraries/version-history
   -- below also sets the GHC version effectively. set to == 8.10.*, == 9.0.*., == 9.2.*, == 9.4.*, == 9.6.*
